Generalize the `cleave` function, add 7.07
It no longer requires a trapezoid and circular cleave, but can take one
of each or two of the same. Anything that implements Cleave.

I also added in changes for 7.07. Because Battle Fury has changed
multiple times, I also made the `bfuryTrap` set up values, including
values for the trapezoidal/conical cleave in 7.0, 7.02 and 7.02
separately. the `bfury` value uses the newest one paried with the old
circular cleave.

There has also been some new output. With the generalized `cleave`
function, I added comparisons between patches (7.00->7.02, 7.02-7.07).

Note that this does mean that the type changed, and requires you to
use `uncurry` with old values, as well as removed the `Maybe` from the
second argument in `cleave`. You can just use the `standardCleave` value
there, as it has an area of 0.